rename strategy name
