+++ src/strategies/gains-vault-staked-balance/index.ts
